fix: ampliacion de las tablas, montos de la cotizacion

+++ database/migrations/2020_09_11_084026_create_cotizacions_table.php
@@ -29,9 +29,9 @@ class CreateCotizacionsTable extends Migration
             $table->string('condicion')->nullable(); 
             $table->string('nota_de_seguimiento')->nullable();
             $table->string('ruta_print_document')->nullable();
-            $table->bigInteger('sub_total')->nullable();
-            $table->bigInteger('iva')->nullable();
-            $table->bigInteger('total')->nullable();
+            $table->float('sub_total', 8, 2)->nullable();
+            $table->float('iva', 8, 2)->nullable();
+            $table->float('total', 8, 2)->nullable();
             $table->timestamps();
             $table->softDeletes()->nullable();
         });
